Separate splash page client logic
